// packages/react-datasheet-grid/src/types.ts
export type CellProps<T, C> = {
  rowData: T
  rowIndex: number
  columnIndex: number
  active: boolean
  focus: boolean
  disabled: boolean
  columnData: C
  setRowData: (rowData: T) => void
  stopEditing: (opts?: { nextRow?: boolean }) => void
  insertRowBelow: () => void
  duplicateRow: () => void
  deleteRow: () => void
  getContextMenuItems: () => ContextMenuItem[]
}

export type CellComponent<T, C> = (props: CellProps<T, C>) => React.JSX.Element

export type Column<T, C, PasteValue> = {
  id?: string
  headerClassName?: string
  title?: React.ReactNode
  /** @deprecated Use `basis`, `grow`, and `shrink` instead */
  width?: string | number
  basis: number
  grow: number
  shrink: number
  minWidth: number
  maxWidth?: number
  component: CellComponent<T, C>
  columnData?: C
  disableKeys: boolean
  disabled: boolean | ((opt: { rowData: T; rowIndex: number }) => boolean)
  cellClassName?:
    | string
    | ((opt: {
        rowData: T
        rowIndex: number
        columnId?: string
      }) => string | undefined)
  keepFocus: boolean
  deleteValue: (opt: { rowData: T; rowIndex: number }) => T
  copyValue: (opt: { rowData: T; rowIndex: number }) => number | string | null
  pasteValue: (opt: { rowData: T; value: PasteValue; rowIndex: number }) => T
  prePasteValues: (values: string[]) => PasteValue[] | Promise<PasteValue[]>
  isCellEmpty: (opt: { rowData: T; rowIndex: number }) => boolean
  stickyLeft?: boolean
}
// ...
export type ContextMenuItem =
  | {
      type:
        | 'INSERT_ROW_BELLOW'
        | 'DELETE_ROW'
        | 'DUPLICATE_ROW'
        | 'COPY'
        | 'CUT'
        | 'PASTE'
      action: () => void
    }
  | {
      type: 'DELETE_ROWS' | 'DUPLICATE_ROWS'
      action: () => void
      fromRow: number
      toRow: number
    }